requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
colorlog==6.7.0
Django==4.1.3
django-log-request-id==2.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
dnspython==2.2.1
PyJWT==2.6.0
pymongo==3.12.3
python-environ==0.4.54
pytz==2022.6
sqlparse==0.2.4
